refactor: move modules into ./modules
